Upgrade userland to Debian Buster
Summary:
The updated clang-format package provided for Buster already creates
the clang-format binary.

The completions for Mercurial and Subversion aren't shipped anymore,
and we don't seem to have a current need for them.

This is a follow-up for 8b10c0ea630b.

Test Plan: Build image, run it on Dwellers.

Reviewers: dereckson

Reviewed By: dereckson

Subscribers: #nasqueron_docker_deployment_squad

Tags: #docker_images

Differential Revision: https://devcentral.nasqueron.org/D2345