Allow to run the arc container as unprivileged user more easily
Summary:
Provide an updated arc wrapper to build by-user images
to be able to run comfortably Arcanist or a shell with
it as an unprivileged user.

As this could be considered an hacky solution, or as it
could be used on machine where `docker build` isn't
available, we keep the legacy one too.

Document both the wrappers.

This change is imported from Nasqueron Operations repository,
sync'ed at commit rOPScfb07164325c.

Test Plan:
Both wrappers have been tested on Dwellers,
as the `arc` command.

Reviewers: dereckson

Reviewed By: dereckson

Subscribers: #nasqueron_docker_deployment_squad

Tags: #docker_images

Differential Revision: https://devcentral.nasqueron.org/D2349