Allow tu update user image
Summary:
The change 64580849cf92 has introduced a way to run the arc container
as unprivileged user more easily, with a mechanism to automatically
build a custom image for each user.

This change provies an `arc update` command to pull the latest image
and rebuild the custom image if needed.

Test Plan: `arc update`

Reviewers: dereckson

Reviewed By: dereckson

Subscribers: #nasqueron_docker_deployment_squad

Tags: #docker_images

Differential Revision: https://devcentral.nasqueron.org/D2370